move account claiming logic from box into a separate inner accountClaimer module
