+++ src/box/box.controller.ts
@@ -45,47 +45,6 @@ export class BoxController {
   ) {}
 
   /**
-   * Claim tester account.
-   *
-   * @remarks Tester can claim his/her account for the testing box.
-   *
-   * Notice that the tester should know the shared testers password in order to be able to clain the account.
-   * This password is available after the group admin has started the session.
-   *
-   * Notice that the endpoint should be called only from browser, since a cookie should be added by the API
-   */
-  @ApiResponseDescription({
-    success: {
-      status: 200,
-      type: ClaimAccountResponseDto,
-    },
-    errors: [400, 403, 404],
-    hasAuth: false,
-  })
-  @NoAuth()
-  @Get('claim-account')
-  @UniformResponse(undefined, ClaimAccountResponseDto)
-  async claimAccount(
-    @Req() req: Request,
-    @Res() res: Response,
-    @Query('password') password: string,
-  ) {
-    const { accountClaimed } = req.cookies;
-    if (accountClaimed)
-      throw new APIError({
-        reason: APIErrorReason.NOT_AUTHORIZED,
-        message: 'Account already claimed from this device.',
-      });
-
-    const data = await this.service.claimAccount(password);
-    res.cookie('accountClaimed', true, {
-      httpOnly: false,
-      maxAge: 15 * 60 * 1000,
-    });
-    return res.send(data);
-  }
-
-  /**
    * Create a testing box.
    *
    * @remarks Create a testing box.
